feat, test: remove disputeWitness, more robust dispute testing
While adding more robust testing, the following issue was discovered with
dispute witnesses. Assume that correct states are [0..19] in increments of 1.
Assume that incorrect states are [0..9, 15.1..19.1] in increments of 1.

The following sequence of calls can occur:
1. Challenge manager is initialized with [0, 9, 19] by the challenger.
Challenger uses correct states.
2. Split is called with consensusWitness of 9, states of [14] and
disputeWitness of 19. disputeWitness needs to be a witness for the merkle
root of states recorded on chain.
3. The challenger does not disagree with any of the states in the challenger
manager.

Moreover, even if above was not an issue,it doesn't seem that the dispute
witness validation adds any value to the dispute game.